# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SichrPlace API",
    "version": "1.0.0",
    "description": "API documentation for SichrPlace backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/messages/{username}": {
      "get": {
        "summary": "Get all messages for a user",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of messages" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
...
}